chatapp/webapp/requirements.txt
Flask==3.0.2
gunicorn==21.2.0
Flask-MySQLdb==2.0.0
Flask-Session==0.6.0
cryptography==42.0.5
PyYAML==6.0.1
Flask-SocketIO==5.3.6
eventlet==0.35.2
Flask-Bcrypt
pycryptodome
onetimepass
pyqrcode
flask_login
wtforms
requests
Pillow